Move NSItemProvider handling into SignalMessaging
There's more I'm planning on changing here, but for now this is mostly
just a copy/paste of code into SignalMessaging. Anything that used to
live in the ShareViewController to handle translating from
NSItemProvider to SignalAttachment now lives in a container enum
"OWSItemProvider".

The enum is just used to ensure there aren't any namespace collisions.